Develop

+++ src/main/java/com/demo/server/common/utils/file/FileUtil.java
@@ -9,10 +9,10 @@ package com.demo.server.common.utils.file;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
+import org.apache.commons.lang3.StringUtils;
 
-import javax.servlet.http.HttpServletRequest;
-import java.io.*;
-import java.net.URLEncoder;
+import java.io.File;
+import java.io.IOException;
 
 /**
  * 文件处理工具类
@@ -22,64 +22,37 @@ import java.net.URLEncoder;
  */
 @Slf4j
 public class FileUtil extends org.apache.commons.io.FileUtils {
-    public static String FILENAME_PATTERN = "[a-zA-Z0-9_\\-\\|\\.\\u4e00-\\u9fa5]+";
-
-    /**
-     * 输出指定文件的byte数组
-     *
-     * @param filePath 文件路径
-     * @param os       输出流
-     * @return
-     */
-    public static void writeBytes(String filePath, OutputStream os) throws IOException {
-        FileInputStream fis = null;
-        try {
-            File file = new File(filePath);
-            if (!file.exists()) {
-                throw new FileNotFoundException(filePath);
-            }
-            fis = new FileInputStream(file);
-            byte[] b = new byte[1024];
-            int length;
-            while ((length = fis.read(b)) > 0) {
-                os.write(b, 0, length);
-            }
-        } catch (IOException e) {
-            log.error("writeBytes error.", e);
-        } finally {
-            if (os != null) {
-                try {
-                    os.close();
-                } catch (IOException e1) {
-                    log.error("writeBytes error.", e1);
-                }
-            }
-            if (fis != null) {
-                try {
-                    fis.close();
-                } catch (IOException e1) {
-                    log.error("writeBytes error.", e1);
-                }
-            }
-        }
-    }
+    public static final String FILENAME_PATTERN = "[a-zA-Z0-9_\\-\\|\\.\\u4e00-\\u9fa5]+";
 
     public static String getExtName(String filename) {
         return FilenameUtils.getExtension(filename);
     }
 
-    public static boolean deleteFile(String filePath) {
+    public static void moveFile(File srcFile, File destFile) throws IOException {
+        FileUtils.moveFile(srcFile, destFile);
+        log.info("moveFile ok. src={},dest={}", srcFile.getAbsolutePath(), destFile.getAbsolutePath());
+    }
+
+    public static boolean deleteFile(String path) {
+        if (StringUtils.isBlank(path)) {
+            log.error("deleteFile error. filePath can not blank.");
+            return false;
+        }
         boolean flag = false;
-        File file = new File(filePath);
-        // 路径为文件且不为空则进行删除
+        File file = new File(path);
         if (file.isFile() && file.exists()) {
             file.delete();
             flag = true;
+            log.info("deleteFile ok. path={}", path);
         }
         return flag;
     }
 
     public static boolean cleanDirectory(String path) {
+        if (StringUtils.isBlank(path)) {
+            log.error("cleanDirectory error. path can not blank.");
+            return false;
+        }
         File file = new File(path);
         if (!file.exists() || !file.isDirectory()) {
             log.error("cleanDirectory error. path={} is not exists or not directory", path);
@@ -95,41 +68,11 @@ public class FileUtil extends org.apache.commons.io.FileUtils {
         return false;
     }
 
-    /**
-     * 文件名称验证
-     *
-     * @param filename 文件名称
-     * @return true 正常 false 非法
-     */
-    public static boolean isValidFilename(String filename) {
-        return filename.matches(FILENAME_PATTERN);
+    public static String getReadableSizeStr(long size) {
+        return FileUtils.byteCountToDisplaySize(size);
     }
 
-    /**
-     * 下载文件名重新编码
-     *
-     * @param request  请求对象
-     * @param fileName 文件名
-     * @return 编码后的文件名
-     */
-    public static String setFileDownloadHeader(HttpServletRequest request, String fileName)
-            throws UnsupportedEncodingException {
-        final String agent = request.getHeader("USER-AGENT");
-        String filename = fileName;
-        if (agent.contains("MSIE")) {
-            // IE浏览器
-            filename = URLEncoder.encode(filename, "utf-8");
-            filename = filename.replace("+", " ");
-        } else if (agent.contains("Firefox")) {
-            // 火狐浏览器
-            filename = new String(fileName.getBytes(), "ISO8859-1");
-        } else if (agent.contains("Chrome")) {
-            // google浏览器
-            filename = URLEncoder.encode(filename, "utf-8");
-        } else {
-            // 其它浏览器
-            filename = URLEncoder.encode(filename, "utf-8");
-        }
-        return filename;
+    public static boolean isValidFilename(String filename) {
+        return filename.matches(FILENAME_PATTERN);
     }
 }
